Fixing error in trivia.html
To fix the "The value of the for attribute of the label element must be the ID of a non-hidden form control" errors on lines 33 and 55, we fixed the for attributes in the input tags above so that they matched the for attributes in the corresponding label tags below. Now the for attribute on line 33="question2.4" instead of "question.4" and the for attribute on line 55="question4.4" instead of "question.4".